2026/clickhouse_llm_integration_knn/requirements.txt
sentence-transformers
clickhouse-connect
numpy
openai
python-dotenv
